COMP: Remove warnings and fix shared build

+++ VTK/Utilities/vtkexodus2/excn2s.c
@@ -57,7 +57,7 @@
 *
 * revision history - 
 *
-*  $Id: excn2s.c,v 1.1 2005-07-17 15:43:59 andy Exp $
+*  $Id: excn2s.c,v 1.2 2005-07-19 12:31:28 andy Exp $
 *
 *****************************************************************************/
 
@@ -350,7 +350,7 @@ int ex_cvt_nodes_to_sides(int exoid,
     elem_blk_parms[i].num_nodes_per_elem = num_nodes_per_elem;
     elem_blk_parms[i].num_attr = num_attr;
 
-    for (m=0; m < strlen(elem_type); m++)
+    for (m=0; (size_t)m < strlen(elem_type); m++)
       elem_blk_parms[i].elem_type[m] = 
               toupper((int)elem_type[m]);
     elem_blk_parms[i].elem_type[m] = '\0';
