+++ VTK/Utilities/vtkexodus2/excre.c
@@ -55,7 +55,7 @@
 *
 * revision history - 
 *
-*  $Id: excre.c,v 1.1 2005-07-17 15:43:59 andy Exp $
+*  $Id: excre.c,v 1.2 2005-07-19 12:31:28 andy Exp $
 *
 *****************************************************************************/
 #include "exodusII.h"
@@ -152,7 +152,7 @@ int ex_create (const char *path,
  */
 
 /* store Exodus API version # as an attribute */
-  vers = EX_API_VERS;
+  vers = (float)(EX_API_VERS);
   if (ncattput (exoid, NC_GLOBAL, ATT_API_VERSION, NC_FLOAT, 1, &vers) == -1)
   {
     exerrval = ncerr;
@@ -164,7 +164,7 @@ int ex_create (const char *path,
   }
 
 /* store Exodus file version # as an attribute */
-  vers = EX_VERS;
+  vers = (float)(EX_VERS);
   if (ncattput (exoid, NC_GLOBAL, ATT_VERSION, NC_FLOAT, 1, &vers) == -1)
   {
     exerrval = ncerr;
