+++ VTK/Utilities/vtkexodus2/exgssc.c
@@ -52,7 +52,7 @@
 *                                       side or face
 * revision history - 
 *
-*  $Id: exgssc.c,v 1.1 2005-07-17 15:43:59 andy Exp $
+*  $Id: exgssc.c,v 1.2 2005-07-19 12:31:28 andy Exp $
 *****************************************************************************/
 
 #include <ctype.h>
@@ -300,7 +300,7 @@ int ex_get_side_set_node_count(int exoid,
     elem_blk_parms[i].num_nodes_per_elem = num_nodes_per_elem;
     elem_blk_parms[i].num_attr = num_attr;
 
-    for (m=0; m < strlen(elem_type); m++) {
+    for (m=0; (size_t)m < strlen(elem_type); m++) {
       elem_blk_parms[i].elem_type[m] = toupper((int)elem_type[m]);
     }
     elem_blk_parms[i].elem_type[m] = '\0';
